Factor out config loader
Main configuration module is doing too much.

Separate automated schema loading from loading individual config options
and misc. special cases.
